Refactor/content poster (#7)
* Remove content poster UI file to split into a UI directory

* Move common UI.py file to its own module/ui directory

* Move content poster UI buttons, embeds, and modals into their respective files in UI directory

* Separate content poster UI views into relevant files in the ui/views directory

* Move common custom made UI from content poster to modules/ui custom file

* Add a typings folder and move content poster types to typings folder

* Add user input functions to a util file

* Update imports in content poster py

* Update imports in modules/twitter py

* Update imports in role picker cogs

* Update imports in orbot py

* Fix OliviaHye hangul hashtag

* Remove unused imports in orbot

* Add comments and clean up code in content poster.py

* Remove buttons.py

* Add comments and clean up code in embeds, modals, persistent py in content poster

* Add comments and clean up code in edit post py

* Add comments and clean up code in new post py

* Add comments and clean up code in post details py

* Add comments and clean up code in src/modules/twitter files

* Add and modify comments in src/modules/ui files

* Add comments and clean up code in src/utils files

* Clean up imports in role picker ui
