+++ src/orbot.py
@@ -4,7 +4,7 @@ import os
 import discord
 from discord.ext import commands
 
-from src.cogs.content_poster.ui import PersistentTweetView
+from src.cogs.content_poster.ui.views.persistent import PersistentTweetView
 from src.cogs.role_picker.ui import PersistentRolePickerView
 from src.modules.twitter.feed import TwitterFeed
 from src.utils.config import ContentPosterConfig
