+++ src/utils/helper.py
@@ -8,7 +8,7 @@ import discord
 
 
 def get_from_dict(dic, map_list):
-    """Iterate nested dictionary"""
+    """Iterate nested dictionary. Returns `None` if not key is not found."""
     try:
         return reduce(dict.get, map_list, dic)
     except TypeError:
@@ -16,10 +16,20 @@ def get_from_dict(dic, map_list):
 
 
 def dict_has_key(dic, key):
+    """Whether or not a dictionary has a given key. Returns `True` or `False`"""
     return key in dic.keys()
 
 
 async def download_files(urls: List[str], filenames: Optional[List[str]] = None):
+    """Downloads multiple files from a list of urls. Returns a list of downloaded `discord.Files`.
+
+    Parameters
+    ----------
+        * urls: List[:class:`str`]
+            - The urls to download.
+        * filenames: Optional[List[`str`]] | None
+            - The filenames for the urls. Must be equal to the number of urls, otherwise it raises an `Exception`.
+    """
     if filenames is not None and len(filenames) != len(urls):
         raise Exception
     return [
@@ -28,6 +38,15 @@ async def download_files(urls: List[str], filenames: Optional[List[str]] = None)
 
 
 async def download_file(url: str, name: str):
+    """Downloads a single file. Returns a downloaded `discord.File` instance.
+
+    Parameters
+    ----------
+        * url: :class:`str`
+            - The url to download.
+        * name: :class:`str`
+            - The name of the downloaded file.
+    """
     async with aiohttp.ClientSession() as session:
         async with session.get(url) as resp:
             if resp.status != 200:
@@ -37,6 +56,15 @@ async def download_file(url: str, name: str):
 
 
 async def convert_files_to_zip(files: List[discord.File], filename: Optional[str] = None):
+    """Converts a list of `discord.File`s to a ZIP file.
+
+    Parameters
+    ----------
+        * files: List[:class:`discord.File`]
+            - The list of files to compress into a ZIP file.
+        * filename: Optional[:class:`str`] | None
+            - The name of the ZIP file.
+    """
     zip_buffer = io.BytesIO()
     with zipfile.ZipFile(zip_buffer, "a", zipfile.ZIP_DEFLATED, False) as zip_file:
         for discord_file in files:
